+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.0.17 - in progress
 - Update atacseq cedar link
 - Add Phenocycler next-gen directory schema
+- Update Histology next-gen directory schema
 
 ## v0.0.16