+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Update atacseq cedar link
 - Add Phenocycler next-gen directory schema
 - Update Histology next-gen directory schema
+- Add LC-MS next-gen directory schema
 
 ## v0.0.16