+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## v0.0.17 - in progress
 - Update atacseq cedar link
+- Add Phenocycler next-gen directory schema
+- Update Histology next-gen directory schema
+- Add LC-MS next-gen directory schema
+- Add GeoMx NGS next-gen directory schema
 
 ## v0.0.16