+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add LC-MS next-gen directory schema
 - Add GeoMx NGS next-gen directory schema
 - Update PhenoCycler and Histology to 2.2.0
+- Update CEDAR links for PhenoCycler & Histology
 
 ## v0.0.16