+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add GeoMx NGS next-gen directory schema
 - Update PhenoCycler and Histology to 2.2.0
 - Update CEDAR links for PhenoCycler & Histology
+- Refactor Upload to avoid validating the same contributors.tsv multiple times / running plugins over files multiple times
 
 ## v0.0.16