+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Update CEDAR links for PhenoCycler & Histology
 - Refactor Upload to avoid validating the same contributors.tsv multiple times / running plugins over files multiple times
 - Add entry for segmentation-mask
+- Add Publication next-gen directory schema
 
 ## v0.0.16