+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Modify directory schema validation such that it takes empty directories into account
 - Add Publication next-gen directory schema
 - Update ATAC/RNA/10X documentation
+- Update Cell Dive documentation
 
 ## v0.0.16