+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - Add entry for segmentation-mask
 - Modify directory schema validation such that it takes empty directories into account
 - Add Publication next-gen directory schema
+- Update ATAC/RNA/10X documentation
+- Update Cell Dive documentation
 
 ## v0.0.16